requirements.txt
groq==0.9.0
streamlit==1.37.0
python-dotenv
